Moved CanvasCard, DeleteShapeButton, ShapeAttributesMenu, Toolbar from components directory to Whiteboard page directory.
