+++ Frontend/src/pages/Whiteboard/Canvas.tsx
