+++ Frontend/src/pages/Whiteboard/DeleteShapeButton.tsx
